Code
R Code to find Outliers i.e Observations that have low probability of being a member of a cluster (i.e. are far away from other members of the clusters) are identified as outliers. OR Clusters that have small populations should possibly be considered outliers. In this aspect, clusters populated with less than 1% of the whole population